feat(rome_console): add html printing to markup
